requirements.txt
flask==3.0.2
flask-cors==4.0.0
requests==2.31.0
pandas==2.2.1
matplotlib==3.8.3
google-generativeai==0.3.2
python-dotenv==1.0.1
gunicorn==21.2.0 